requirements.txt
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
bitsandbytes>=0.41.0
accelerate>=0.20.0 